Generalized fill_template to format_template
Applied to Level3 associations
