+++ jwst/associations/association.py
@@ -7,7 +7,6 @@ import jsonschema
 import logging
 from nose.tools import nottest
 import re
-from string import Formatter
 
 from astropy.extern import six
 from numpy.ma import masked
@@ -16,6 +15,7 @@ from . import __version__
 from .exceptions import (
     AssociationNotValidError
 )
+from .lib.format_template import FormatTemplate
 from .lib.ioregistry import IORegistry
 
 __all__ = ['Association']
@@ -273,15 +273,15 @@ class Association(MutableMapping):
         """
         if format is None:
             formats = [
-                format
-                for format_name, format in cls.ioregistry.items()
+                format_func
+                for format_name, format_func in cls.ioregistry.items()
             ]
         else:
             formats = [cls.ioregistry[format]]
 
-        for format in formats:
+        for format_func in formats:
             try:
-                asn = format.load(
+                asn = format_func.load(
                     cls, serialized, **kwargs
                 )
             except AssociationNotValidError:
@@ -702,131 +702,5 @@ def is_iterable(obj):
         hasattr(obj, '__iter__')
 
 
-class FillTemplate(Formatter):
-    """Fill a template
-
-    Parameters
-    ----------
-    template: str
-        A Python format string
-
-    separator: string
-        Separater to use for values which have no
-        matching replacement strings
-
-    kwargs: dict or named parameters
-        The key/values pairs to fill into the Python format string
-
-    Returns
-    -------
-    str
-        The formatted string
-
-    Notes
-    -----
-    This differences from Pythons `format` method are:
-        - If a replacement field does not have a given value,
-          the replacement field is left in the result
-        - If a key/value pair is present but has no replacement field,
-          the value is simply appended.
-        - Template can only use named replacement fields.
-
-    Examples
-    --------
-    The basic example:
-    >>> template = 'name="{name}" value="{value}"'
-    >>> fmt = FillTemplate()
-    >>> fmt(template)
-    'name="{name}" value="{value}"'
-
-    But with actual values given:
-    >>> fmt(template, name='fred', value='great')
-    'name="fred" value="great"'
-
-    But wait, too many values given:
-    >>> fmt(template, name='fred', value='great', extra='more')
-    'name="fred" value="great"_more'
-
-    And with too many and not enough:
-    >>> fmt(template, value='great', extra='more')
-    'name="{name}" value="great"_more'
-
-    With a different separator:
-    >>> fmt.separator = '---'
-    >>> fmt(template, name='fred', value='great', extra='more')
-    'name="fred" value="great"---more'
-
-    Initializing with a different separator:
-    >>> fmt_newsep = FillTemplate(separator='_now-with_')
-    >>> fmt_newsep(template, name='fred', value='great', extra='more')
-    'name="fred" value="great"_now-with_more'
-    """
-    def __init__(self, separator='_'):
-        """Inialize class
-
-        Parameters
-        ----------
-        separator: str
-            For key/value pairs given that do not have a
-            replacement field, the values are appened to
-            the string using this separator.
-        """
-        super(FillTemplate, self).__init__()
-        self.separator = separator
-
-    def format(self, format_string, **kwargs):
-        """Perform the formatting
-
-        Parameters
-        ----------
-        format_string: str
-            The string to be formatted
-
-        kwargs: dict
-            The key/value pairs to insert into the string
-
-        Returns
-        -------
-        formatted: str
-            The formatted string.
-        """
-        self._used_keys = []
-        result = super(FillTemplate, self).format(format_string, **kwargs)
-
-        # Get any unused arguments and simply do the appending
-        unused_keys = set(kwargs).difference(self._used_keys)
-        unused_values = [kwargs[unused] for unused in unused_keys]
-        result_parts = [result] + unused_values
-        result = self.separator.join(result_parts)
-
-        return result
-    __call__ = format
-
-    def get_value(self, key, args, kwargs):
-        """Return a given field value
-
-        Parameters
-        ----------
-        key: str
-            The key to retrieve.
-
-        args: [arg(, ...)]
-            Positional arguments passed.
-            This is ignored.
-
-        kwargs: {k:v(, ...)}
-            The key/value pairs passed in.
-
-        Returns
-        -------
-        obj
-            The value from the kwargs.
-            If not found, the string '{key}' is returned.
-        """
-        value = kwargs.get(key, '{' + key + '}')
-        self._used_keys.append(key)
-
-        return value
-
-# Define the default utility function for FillTemplate
-fill_template = FillTemplate()
+# Define default product name filling
+format_product = FormatTemplate()
